requirements.txt
streamlit_option_menu==0.3.13
py3Dmol==2.4.0
stmol==0.0.9
ipython_genutils==0.2.0